backend/requirement.txt
asgiref==3.4.1
certifi==2022.12.7
charset-normalizer==2.0.12
Django==3.2
django-ckeditor==6.3.0
django-cors-headers==3.13.0
django-crispy-forms==1.13.0
django-js-asset==2.0.0
django-utils-six==2.0
djangorestframework==3.14.0
idna==3.4
Pillow==8.4.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2022.4
Unidecode==1.3.6
urllib3==1.26.14
whitenoise==5.3.0
